Combine LifecycleTests and BaseLifecycleTests, now that there is only one inheritor.
