+++ src/Fixie.Tests/LifecycleTests.cs
@@ -1,14 +1,128 @@
-﻿namespace Fixie.Tests.Lifecycle
+﻿namespace Fixie.Tests
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Reflection;
+    using System.Runtime.CompilerServices;
     using Assertions;
-    using Lifecycle = Fixie.Lifecycle;
+    using Fixie.Execution;
     using static System.Environment;
 
-    public class LifecycleTests : BaseLifecycleTests
+    public class LifecycleTests
     {
+        static string[] FailingMembers;
+        readonly Convention Convention;
+
+        public LifecycleTests()
+        {
+            FailingMembers = null;
+
+            Convention = new Convention();
+            Convention.Classes.Where(testClass => testClass == typeof(SampleTestClass));
+            Convention.Methods.Where(method => method.Name == "Pass" || method.Name == "Fail");
+            Convention.ClassExecution.SortCases((x, y) => String.Compare(y.Name, x.Name, StringComparison.Ordinal));
+        }
+
+        static void FailDuring(params string[] failingMemberNames)
+        {
+            FailingMembers = failingMemberNames;
+        }
+
+        Output Run()
+        {
+            var listener = new StubListener();
+
+            using (var console = new RedirectedConsole())
+            {
+                Utility.Run<SampleTestClass>(listener, Convention);
+
+                return new Output(console.Lines().ToArray(), listener.Entries.ToArray());
+            }
+        }
+
+        class Output
+        {
+            readonly string[] lifecycle;
+            readonly string[] results;
+
+            public Output(string[] lifecycle, string[] results)
+            {
+                this.lifecycle = lifecycle;
+                this.results = results;
+            }
+
+            public void ShouldHaveLifecycle(params string[] expected)
+            {
+                lifecycle.ShouldEqual(expected);
+            }
+
+            public void ShouldHaveResults(params string[] expected)
+            {
+                var namespaceQualifiedExpectation = expected.Select(x => "Fixie.Tests.LifecycleTests+" + x).ToArray();
+
+                results.ShouldEqual(namespaceQualifiedExpectation);
+            }
+        }
+
+        class SampleTestClass : IDisposable
+        {
+            bool disposed;
+
+            public SampleTestClass()
+            {
+                WhereAmI();
+            }
+
+            public void SetUpA()
+            {
+                WhereAmI();
+            }
+
+            public void SetUpB()
+            {
+                WhereAmI();
+            }
+
+            public void Pass()
+            {
+                WhereAmI();
+            }
+
+            public void Fail()
+            {
+                WhereAmI();
+                throw new FailureException();
+            }
+
+            public void TearDownA()
+            {
+                WhereAmI();
+            }
+
+            public void TearDownB()
+            {
+                WhereAmI();
+            }
+
+            public void Dispose()
+            {
+                if (disposed)
+                    throw new ShouldBeUnreachableException();
+                disposed = true;
+
+                WhereAmI();
+            }
+        }
+
+        static void WhereAmI([CallerMemberName] string member = null)
+        {
+            Console.WriteLine(member);
+
+            if (FailingMembers != null && FailingMembers.Contains(member))
+                throw new FailureException(member);
+        }
+
         static void ClassSetUp(Type testClass)
         {
             testClass.ShouldEqual(typeof(SampleTestClass));
